More thorough backwards compatibility for http-client and base
To achieve backwards compatibility for http-client<0.4.30
parseURL needs to be used and the default `checkStatus` handler
overridden to disable the trapping of non-2xx HTTP response codes.
Thereafter `parseRequest` can be used which does not trap any
reponse codes (the required behaviour).

For `base`<4.8, the `Applicative` class is not in the prelude
and must be imported explicitly.